requirements.txt
aiohttp==2.0.2
async-timeout==1.2.0
certifi==2017.1.23
chardet==2.3.0
future==0.16.0
multidict==2.1.4
python-telegram-bot==5.3.0
urllib3==1.20
yarl==0.10.0
